pyproject.toml
[project]
name = "DOC"
version = "2023.11.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    ]
# We use pip-tools to generate requirements.txt files for pinned
# dependences, but this stanza is for other projects relying on DOC as
# a package to take advantage of it as a library.
dependencies = [
  "aiofiles",
  "celery",
  "celery-types",
  "docxtpl",
  "fastapi[all]",
  "flower",
  "htmldocx",
  "html2docx",
  "gunicorn",
  "jinja2",
  "orjson",
  "pydantic",
  "email-validator",
  "python-dotenv",
  "pyyaml",
  "redis",
  "requests",
  "termcolor",
  "toolz",
  "uvicorn",
  "weasyprint",

  "types-PyYAML",
  "types-Markdown",
  "types-beautifulsoup4",
  "types-ujson",
  "types-orjson",
  "types-requests",
  "types-termcolor",
  "types-toml",
  "types-setuptools",
]
[build-system]
requires = [
    "setuptools>=61",
    "wheel",
    "cython",
    "mypy",
]
build-backend = "setuptools.build_meta"
[tool.setuptools.packages.find]
where = ["backend", "tests"]
[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
log_cli = true
addopts = [
    "--tb=long",
    "--showlocals",
    # "-x",
#   "--pdb",
#   "--maxfail=4",
#   "--last-failed",
    "--failed-first",
    # "--log-file=tests_output.log",
    "--log-cli-level=DEBUG",
    "--strict-markers",
    ]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::SyntaxWarning",
    ]
markers = [
    "datafiles", # : Load datafiles fixture
    "slow", # : Marks tests as slow (deselect with '-m "not slow"')
    "randomized", # : Tests that are randomly built combinatorically at runtime using fixtures and not suitable for ci/cd production runs
    "english_randomized", # Tests that are randomly selected
    # combinatorically for English at runtime using fixtures and not suitable for
    # ci/cd production runs
    "docx", # : Marks the majority of tests that test generation of docx output so that they can be run in isolation during development.
    "all", # : Marks a single test which exercises potentially all languages.
    "focus", # : Marks a subset of tests which we want to run instead
    # of the whole suite
    ]
# Speed up pytest collection times
norecursedirs = [
  "docs",
  "*.egg-info",
  ".git",
]
[tool.isort]
profile = "black"
multi_line_output = 3
